# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "auth.ericminassian.com",
    "description": "Personal OIDC provider — SPA API and token endpoint.",
    "license": {
      "name": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/signup/start": {
      "post": {
        "tags": [
          "signup"
        ],
        "summary": "POST /api/signup/start — sends an OTP (or an \"account exists\" notice).\nUniform 200 regardless of account existence (anti-enumeration).",
        "operationId": "start",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OTP dispatched (uniform regardless of account existence)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/signup/verify": {
      "post": {
        "tags": [
          "signup"
        ],
        "summary": "POST /api/signup/verify — consumes the OTP, creates the account, and\nissues an enroll-level session whose only capability is registering the\nfirst passkey.",
        "operationId": "verify",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Account created; enroll-level session cookie set",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerifyResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid or expired code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Account already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/webauthn/login/finish": {
      "post": {
        "tags": [
          "webauthn"
        ],
        "summary": "POST /api/webauthn/login/finish — verifies the assertion and issues a\nfull session.",
        "description": "Body: `{ ceremony_id, credential }` where `credential` is the\n`navigator.credentials.get()` result. Returns `{ user_id }`.",
        "operationId": "login_finish",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginFinishRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "{ user_id }"
          },
          "401": {
            "description": "Authentication failed"
          }
        }
      }
    },
    "/api/webauthn/login/start": {
      "post": {
        "tags": [
          "webauthn"
        ],
        "summary": "POST /api/webauthn/login/start — no prior auth. Without an email this is\nthe discoverable (conditional-UI) flow; with one, an allowCredentials\nchallenge for that account. Unknown emails get a discoverable challenge\nso the response shape stays uniform.",
        "description": "Body: `{ email? }`. Returns `{ ceremony_id, options }` for\n`navigator.credentials.get()`.",
        "operationId": "login_start",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginStartRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "{ ceremony_id, options }"
          }
        }
      }
    },
    "/api/webauthn/register/finish": {
      "post": {
        "tags": [
          "webauthn"
        ],
        "summary": "POST /api/webauthn/register/finish — stores the passkey; upgrades an\nenroll session to full when it was the first registration step.",
        "description": "Body: `{ ceremony_id, credential, name? }` where `credential` is the\n`navigator.credentials.create()` result.",
        "operationId": "register_finish",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterFinishRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "{ credential_id }"
          },
          "400": {
            "description": ""
          },
          "403": {
            "description": ""
          }
        }
      }
    },
    "/api/webauthn/register/start": {
      "post": {
        "tags": [
          "webauthn"
        ],
        "summary": "POST /api/webauthn/register/start — enroll or full session.",
        "description": "Returns `{ ceremony_id, options }` where `options` is the raw\n`PublicKeyCredentialCreationOptions` to pass to `navigator.credentials\n.create()`. The shape is the standard WebAuthn structure (not modeled in\nthis schema).",
        "operationId": "register_start",
        "responses": {
          "200": {
            "description": "{ ceremony_id, options }"
          },
          "401": {
            "description": ""
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "description": "Error envelope returned by every `/api/*` failure.",
        "required": [
          "error",
          "message"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Machine-readable error code, e.g. `rate_limited`, `account_exists`."
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "LoginFinishRequest": {
        "type": "object",
        "required": [
          "ceremony_id",
          "credential"
        ],
        "properties": {
          "ceremony_id": {
            "type": "string"
          },
          "credential": {
            "type": "object",
            "description": "Raw `navigator.credentials.get()` result (standard WebAuthn shape)."
          }
        }
      },
      "LoginStartRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "OkResponse": {
        "type": "object",
        "description": "Generic success envelope (`{ \"ok\": true }`).",
        "required": [
          "ok"
        ],
        "properties": {
          "ok": {
            "type": "boolean"
          }
        }
      },
...
      "RegisterFinishRequest": {
        "type": "object",
        "required": [
          "ceremony_id",
          "credential"
        ],
        "properties": {
          "ceremony_id": {
            "type": "string"
          },
          "credential": {
            "type": "object",
            "description": "Raw `navigator.credentials.create()` result (standard WebAuthn shape)."
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "StartRequest": {
        "type": "object",
        "required": [
          "email"
        ],
        "properties": {
          "email": {
            "type": "string"
          }
        }
      },
...
      "VerifyRequest": {
        "type": "object",
        "required": [
          "email",
          "code"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "email": {
            "type": "string"
          }
        }
      },
      "VerifyResponse": {
        "type": "object",
        "required": [
          "user_id"
        ],
        "properties": {
          "user_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      }
    }
  },
...
}